+++ packages/core/modules/import/index.js
@@ -1,7 +1,7 @@
 export {
-  getTree, loadTree, isValidTree,
-  // deprecated
-  checkTree,
+  getTree, loadTree,
+  // for backward compatibility:
+  checkTree, isValidTree,
   // candidates for moving to utils
   isImmutableTree, isTree, isJsonLogic, jsToImmutable
 } from "./tree";